Final Project/Code/requirements.txt
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
idna==2.9
musicbrainzngs==0.7.1
networkx==2.4
numpy==1.18.1
plotly==4.5.4
python-dateutil==2.8.1
requests==2.23.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
